+++ src/main/java/org/spout/vanilla/api/event/world/WeatherChangeEvent.java
@@ -24,49 +24,53 @@
  * License and see <http://spout.in/licensev1> for the full license, including
  * the MIT license.
  */
-package org.spout.vanilla.plugin.event.world;
+package org.spout.vanilla.api.event.world;
 
-import org.spout.api.event.Event;
+import org.spout.api.event.Cancellable;
 import org.spout.api.event.HandlerList;
-import org.spout.api.geo.discrete.Point;
+import org.spout.api.event.world.WorldEvent;
+import org.spout.api.geo.World;
 import org.spout.api.protocol.event.ProtocolEvent;
 
-import org.spout.vanilla.plugin.data.effect.type.ExplosionEffect;
+import org.spout.vanilla.api.data.Weather;
 
-public class PlayExplosionEffectEvent extends Event implements ProtocolEvent {
+public class WeatherChangeEvent extends WorldEvent implements Cancellable, ProtocolEvent {
 	private static HandlerList handlers = new HandlerList();
-	private Point position;
-	private ExplosionEffect effect;
-	private float size;
+	private Weather current, weather;
 
-	public PlayExplosionEffectEvent(Point position, ExplosionEffect effect, float size) {
-		this.position = position;
-		this.effect = effect;
-		this.size = size;
+	public WeatherChangeEvent(World world, Weather current, Weather weather) {
+		super(world);
+		this.current = current;
+		this.weather = weather;
 	}
 
 	/**
-	 * Gets the Position where the Sound should be played
-	 * @return position of the Sound
+	 * Gets the weather at the time the event is called.
+	 * @return the current weather.
 	 */
-	public Point getPosition() {
-		return this.position;
+	public Weather getCurrentWeather() {
+		return current;
 	}
 
 	/**
-	 * Gets the Effect to play
-	 * @return the Effect
+	 * Gets the new weather set after the event.
+	 * @return the new weather.
 	 */
-	public ExplosionEffect getEffect() {
-		return this.effect;
+	public Weather getNewWeather() {
+		return weather;
 	}
 
 	/**
-	 * Gets the size of the Explosion Effect
-	 * @return Effect size
+	 * Sets the outcome of the event.
+	 * @param weather
 	 */
-	public float getSize() {
-		return this.size;
+	public void setNewWeather(Weather weather) {
+		this.weather = weather;
+	}
+
+	@Override
+	public void setCancelled(boolean cancelled) {
+		super.setCancelled(cancelled);
 	}
 
 	@Override
