+++ src/main/java/org/spout/vanilla/plugin/data/effect/store/GeneralEffects.java
@@ -28,12 +28,12 @@ package org.spout.vanilla.plugin.data.effect.store;
 
 import org.spout.vanilla.api.data.effect.Effect;
 
-import org.spout.vanilla.plugin.data.effect.BatchEffect;
-import org.spout.vanilla.plugin.data.effect.GeneralEffect;
-import org.spout.vanilla.plugin.data.effect.SoundEffect;
+import org.spout.vanilla.api.data.effect.BatchEffect;
+import org.spout.vanilla.api.data.effect.GeneralEffect;
+import org.spout.vanilla.api.data.effect.SoundEffect;
 import org.spout.vanilla.plugin.data.effect.type.BatchExplosionEffect;
 import org.spout.vanilla.plugin.data.effect.type.BreakBlockEffect;
-import org.spout.vanilla.plugin.data.effect.type.ExplosionEffect;
+import org.spout.vanilla.api.data.effect.ExplosionEffect;
 import org.spout.vanilla.plugin.data.effect.type.LavaFizzEffect;
 import org.spout.vanilla.plugin.data.effect.type.MusicDiscEffect;
 import org.spout.vanilla.plugin.data.effect.type.NoteParticleEffect;
