+++ src/main/java/org/spout/vanilla/plugin/protocol/VanillaNetworkSynchronizer.java
@@ -106,11 +106,11 @@ import org.spout.vanilla.plugin.event.window.WindowItemsEvent;
 import org.spout.vanilla.plugin.event.window.WindowOpenEvent;
 import org.spout.vanilla.plugin.event.window.WindowPropertyEvent;
 import org.spout.vanilla.plugin.event.window.WindowSlotEvent;
-import org.spout.vanilla.plugin.event.world.PlayExplosionEffectEvent;
-import org.spout.vanilla.plugin.event.world.PlayParticleEffectEvent;
-import org.spout.vanilla.plugin.event.world.PlaySoundEffectEvent;
-import org.spout.vanilla.plugin.event.world.TimeUpdateEvent;
-import org.spout.vanilla.plugin.event.world.WeatherChangeEvent;
+import org.spout.vanilla.api.event.world.PlayExplosionEffectEvent;
+import org.spout.vanilla.api.event.world.PlayParticleEffectEvent;
+import org.spout.vanilla.api.event.world.PlaySoundEffectEvent;
+import org.spout.vanilla.api.event.world.TimeUpdateEvent;
+import org.spout.vanilla.api.event.world.WeatherChangeEvent;
 import org.spout.vanilla.plugin.inventory.window.DefaultWindow;
 import org.spout.vanilla.plugin.material.VanillaMaterials;
 import org.spout.vanilla.plugin.protocol.container.VanillaContainer;
